Uploading first small project - Magic Ball
The Magic 8 Ball is a plastic sphere, made to look like an oversized eight ball, that is used for fortune-telling or seeking advice. It was invented in 1946 by Albert C. Carter and Abe Bookman and is currently manufactured by Mattel. The user asks a yes–no question to the ball, then turns it over to reveal an answer that floats up into a window. (wiki)